Moved Torch & New Guard Created created.
A new guard has been created and named the Elite Guard. The left cell was converted into a guard breakout room that holds the key to the locked door. A nerfed elite guard is patrolling that area. A new flashlight has been added to the game, right next the converted room.